refactor: space.type.ts 통합
