+++ frontend/src/types/domain/space.type.ts
@@ -1,3 +1,13 @@
+export interface MySpace {
+  id: number;
+  spaceCode: string;
+  title: string;
+  thumbnail: string;
+  guestCount: number;
+  createdAt: string;
+  isPublic: boolean;
+}
+
 export interface SpacePhoto {
   isExists: boolean;
   path: string;
