+++ frontend/src/types/space.type.ts
@@ -1,12 +1,2 @@
-// TODO: PR별 선언 위치 통합 필요
+// TODO: Visibility 형태 변경에 따른 제거 필요(나중에 퍼널과 함께 손대겠음)
 export type SpaceVisibility = 'PUBLIC' | 'PRIVATE';
-
-export interface MySpace {
-  id: number;
-  spaceCode: string;
-  title: string;
-  thumbnail: string;
-  guestCount: number;
-  createdAt: string;
-  isPublic: boolean;
-}
